--- a/backend/src/meal_shield/scrape/cookpad.py
+++ b/backend/src/meal_shield/scrape/cookpad.py
@@ -1,6 +1,6 @@
 from typing import Optional, Union
-
 import aiohttp
+import asyncio
 import requests
 from bs4 import BeautifulSoup
 from tenacity import retry, stop_after_attempt, wait_fixed
@@ -9,7 +9,9 @@ from tqdm.asyncio import tqdm
 # 検索上限(page数)
 LIMIT_PAGE = 35
 MAX_RECIPE_SIZE = 100
+SEMAPHORE_LIMIT = 10  # 同時実行数の制限
 
+semaphore = asyncio.Semaphore(SEMAPHORE_LIMIT)
 
 @retry(stop=stop_after_attempt(1), wait=wait_fixed(1), reraise=True)
 async def scraping_cookpad(
@@ -43,7 +45,6 @@ async def scraping_cookpad(
     else:
         return recipes_list[:MAX_RECIPE_SIZE]
 
-
 @retry(stop=stop_after_attempt(1), wait=wait_fixed(1), reraise=True)
 def make_url_list(recipe_name: str) -> Optional[list[str]]:
     try:
@@ -70,72 +71,72 @@ def make_url_list(recipe_name: str) -> Optional[list[str]]:
     except Exception as e:
         return None
 
-
 @retry(stop=stop_after_attempt(1), wait=wait_fixed(1), reraise=True)
 async def scraping_recipe_url(
     session: aiohttp.ClientSession, url: str
 ) -> Optional[list[str]]:
-    try:
-        async with session.get(url) as response:
-            response.raise_for_status()
-            content = await response.text()
-            soup = BeautifulSoup(content, 'lxml')
+    async with semaphore:
+        try:
+            async with session.get(url) as response:
+                response.raise_for_status()
+                content = await response.text()
+                soup = BeautifulSoup(content, 'lxml')
 
-            recipe_url_list = []
-            # レシピのURLを属性に持つ<a>タグをすべて取得
-            a_tags = soup.find_all('a', class_='recipe-title')
-            for a_tag in a_tags:
-                # 各<a>タグのhref属性(URL)を取得
-                href = a_tag.get('href')
-                # 会員登録が必要なレシピを除外
-                if 'dining' not in href:
-                    recipe_url_list.append(f'https://cookpad.com{href}')
-            return recipe_url_list
-    except Exception as e:
-        return None
-
+                recipe_url_list = []
+                # レシピのURLを属性に持つ<a>タグをすべて取得
+                a_tags = soup.find_all('a', class_='recipe-title')
+                for a_tag in a_tags:
+                    # 各<a>タグのhref属性(URL)を取得
+                    href = a_tag.get('href')
+                    # 会員登録が必要なレシピを除外
+                    if 'dining' not in href:
+                        recipe_url_list.append(f'https://cookpad.com{href}')
+                return recipe_url_list
+        except Exception as e:
+            return None
 
 @retry(stop=stop_after_attempt(1), wait=wait_fixed(1), reraise=True)
 async def scraping_recipe_data(
     session: aiohttp.ClientSession, url: str
 ) -> Optional[dict[str, Union[str, list[str]]]]:
-    try:
-        async with session.get(url) as response:
-            response.raise_for_status()
-            content = await response.text()
+    async with semaphore:
+        try:
+            async with session.get(url) as response:
+                response.raise_for_status()
+                content = await response.text()
 
-            soup = BeautifulSoup(content, 'lxml')
+                soup = BeautifulSoup(content, 'lxml')
 
-            # レシピのタイトルである<h1>タグのテキストを取得
-            recipe_title = soup.find('h1', class_='recipe-title').get_text(strip=True)
-            # 全角スペースを半角スペースに置き換える
-            recipe_title = recipe_title.replace('\u3000', ' ')
+                # レシピのタイトルである<h1>タグのテキストを取得
+                recipe_title = soup.find('h1', class_='recipe-title').get_text(strip=True)
+                # 全角スペースを半角スペースに置き換える
+                recipe_title = recipe_title.replace('\u3000', ' ')
 
-            recipe_ingredients = []
-            # 材料のデータを含む<span>タグの要素をすべて取得
-            spans = soup.find_all('span', class_='name')
-            for span in spans:
-                # 材料は<span>タグかその中の<a>タグにあるので場合分け
-                a_tag = span.find('a')
-                if a_tag is None:
-                    ingredient_name = span.text
-                else:
-                    ingredient_name = a_tag.text
-                recipe_ingredients.append(ingredient_name)
+                recipe_ingredients = []
+                # 材料のデータを含む<span>タグの要素をすべて取得
+                spans = soup.find_all('span', class_='name')
+                for span in spans:
+                    # 材料は<span>タグかその中の<a>タグにあるので場合分け
+                    a_tag = span.find('a')
+                    if a_tag is None:
+                        ingredient_name = span.text
+                    else:
+                        ingredient_name = a_tag.text
+                    recipe_ingredients.append(ingredient_name)
 
-            # レシピ画像のURLを属性に持つ<img>タグを含む<section>タグを取得
-            section_tag = soup.find('section', id='main-photo')
-            # レシピ画像のURLを属性に持つ<img>タグを取得
-            img_tag = section_tag.find('img')
-            # <img>タグのsrc属性(レシピ画像のURL)を取得
-            recipe_image_url = img_tag.get('src')
+                # レシピ画像のURLを属性に持つ<img>タグを含む<section>タグを取得
+                section_tag = soup.find('section', id='main-photo')
+                # レシピ画像のURLを属性に持つ<img>タグを取得
+                img_tag = section_tag.find('img')
+                # <img>タグのsrc属性(レシピ画像のURL)を取得
+                recipe_image_url = img_tag.get('src')
 
-            recipe_data = {
-                'recipe_title': recipe_title,
-                'recipe_ingredients': recipe_ingredients,
-                'recipe_url': url,
-                'recipe_image_url': recipe_image_url,
-            }
-            return recipe_data
-    except Exception as e:
-        return None
+                recipe_data = {
+                    'recipe_title': recipe_title,
+                    'recipe_ingredients': recipe_ingredients,
+                    'recipe_url': url,
+                    'recipe_image_url': recipe_image_url,
+                }
+                return recipe_data
+        except Exception as e:
+            return None