--- a/backend/src/meal_shield/scrape/cookpad.py
+++ b/backend/src/meal_shield/scrape/cookpad.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import Optional, Union, Final
 import aiohttp
 import asyncio
 import requests
@@ -7,11 +7,11 @@ from tenacity import retry, stop_after_attempt, wait_fixed
 from tqdm.asyncio import tqdm
 
 # 検索上限(page数)
-LIMIT_PAGE = 35
-MAX_RECIPE_SIZE = 100
-SEMAPHORE_LIMIT = 10  # 同時実行数の制限
+LIMIT_PAGE: Final[int] = 35
+MAX_RECIPE_SIZE: Final[int] = 100
+SEMAPHORE_LIMIT: Final[int] = 10
 
-semaphore = asyncio.Semaphore(SEMAPHORE_LIMIT)
+semaphore: asyncio.Semaphore = asyncio.Semaphore(SEMAPHORE_LIMIT)
 
 @retry(stop=stop_after_attempt(1), wait=wait_fixed(1), reraise=True)
 async def scraping_cookpad(