--- a/backend/src/meal_shield/scrape/cookpad.py
+++ b/backend/src/meal_shield/scrape/cookpad.py
@@ -8,9 +8,9 @@ from tenacity import retry, stop_after_attempt, wait_fixed
 from tqdm.asyncio import tqdm
 
 # 検索上限(page数)
-LIMIT_PAGE: Final[int] = 35
+LIMIT_PAGE: Final[int] = 40
 MAX_RECIPE_SIZE: Final[int] = 100
-SEMAPHORE_LIMIT: Final[int] = 10
+SEMAPHORE_LIMIT: Final[int] = 1000
 
 semaphore: asyncio.Semaphore = asyncio.Semaphore(SEMAPHORE_LIMIT)