--- a/frontend/src/meal_shield/search.py
+++ b/frontend/src/meal_shield/search.py
@@ -9,8 +9,9 @@ from meal_shield.env import PACKAGE_DIR
 base_url = 'http://backend:8000'
 
 
+BASE_IMAGE_URL = 'https://raw.githubusercontent.com/al22-1B-01/meal_shield/main/frontend/data/images/'
 # アレルギー品目の選択肢
-ALLERGY_OPTION = [
+_ALLERGY_OPTION = [
     {'name': 'えび', 'file': 'ebi.png'},
     {'name': 'かに', 'file': 'kani.png'},
     {'name': 'いか', 'file': 'ika.png'},
@@ -40,6 +41,7 @@ ALLERGY_OPTION = [
     {'name': 'りんご', 'file': 'ringo.png'},
     {'name': 'キウイフルーツ', 'file': 'kiwi.png'},
 ]
+ALLERGY_OPTION = [{'name': item['name'], 'file': BASE_IMAGE_URL + item['file']} for item in _ALLERGY_OPTION]
 
 
 def fetch_recipe_detail(recipe_name: str, allergies: list[str]) -> list:
@@ -75,7 +77,7 @@ def search_recipe_entrypoint() -> None:
         )
 
         col = cols[index % 7]
-        image = Image.open(PACKAGE_DIR / f'data/images/{item["file"]}')
+        # image = Image.open(PACKAGE_DIR / f'data/images/{item["file"]}')
 
         if col.button(item['name']):
             if item['name'] in st.session_state.allergy_list:
@@ -83,7 +85,8 @@ def search_recipe_entrypoint() -> None:
             else:
                 st.session_state.allergy_list.append(item['name'])
 
-        col.image(image, use_column_width=True, width=100)
+        # col.image(image, use_column_width=True, width=100)
+        col.image(item['file'], use_column_width=True, width=100)
 
     st.subheader('選択されたアレルギー品目')
     for allergy in st.session_state.allergy_list: