# WHY
- appの挙動が異常に悪いため
# WHAT
-  アレルギー画像をwebから取得
- スクレイピングのworker数を増加

# Result
- アレルギー画像 -> 目で見て早い
- スクレイピング -> 処理件数次第

実施前
```
Fetching recipe_urls: 100% 35/35 [00:02<00:00, 12.81it/s]
Fetching recipe_data: 100% 350/350 [00:16<00:00, 21.57it/s]
Processing recipes by ChatGPT:  87% 87/100 [00:01<00:00, 65.76it/s]ERROR:meal_shield.ranking.ranking_chatgpt:Failed to parse score: 'NoneType' object has no attribute 'group'
Processing recipes by ChatGPT: 100% 100/100 [00:02<00:00, 33.77it/s]
Processing recipes by embedding: 100% 100/100 [00:02<00:00, 36.39it/s]
Processing recipes by counting: 100% 100/100 [00:00<00:00, 163.92it/s]
```
実施後(worker=10, N_page=35)
```
Fetching recipe_urls: 100% 35/35 [00:03<00:00, 11.06it/s]
Fetching recipe_data: 100% 350/350 [00:18<00:00, 19.21it/s]
Processing recipes by ChatGPT: 100% 98/98 [00:04<00:00, 22.29it/s]
Processing recipes by embedding: 100% 98/98 [00:01<00:00, 91.30it/s]
Processing recipes by counting: 100% 98/98 [00:00<00:00, 166.31it/s]
```
実施後(worker=100, N_page=35)
```
Fetching recipe_urls: 100% 35/35 [00:02<00:00, 15.34it/s]
Fetching recipe_data: 100% 350/350 [00:16<00:00, 21.14it/s]
Processing recipes by ChatGPT:  53% 52/99 [00:01<00:00, 66.88it/s]ERROR:meal_shield.ranking.ranking_chatgpt:Failed to parse score: 'NoneType' object has no attribute 'group'
Processing recipes by ChatGPT:  88% 87/99 [00:01<00:00, 53.73it/s]ERROR:meal_shield.ranking.ranking_chatgpt:Failed to parse score: 'NoneType' object has no attribute 'group'
Processing recipes by ChatGPT: 100% 99/99 [00:02<00:00, 33.93it/s]
Processing recipes by embedding: 100% 99/99 [00:01<00:00, 78.23it/s]
Processing recipes by counting: 100% 99/99 [00:00<00:00, 200.73it/s]
```
実施後(worker=1000, N_page=35)
```
Fetching recipe_urls: 100% 35/35 [00:02<00:00, 15.50it/s]
Fetching recipe_data: 100% 350/350 [00:16<00:00, 20.94it/s]
Processing recipes by ChatGPT: 100% 96/96 [00:02<00:00, 33.76it/s]
Processing recipes by embedding: 100% 96/96 [00:02<00:00, 47.16it/s]
Processing recipes by counting: 100% 96/96 [00:00<00:00, 173.30it/s]
```
実施後(worker=1000, N_page=100)
```
Fetching recipe_urls: 100% 100/100 [00:05<00:00, 16.85it/s]
Fetching recipe_data: 100% 1000/1000 [00:48<00:00, 20.68it/s]
Processing recipes by ChatGPT: 100% 98/98 [00:02<00:00, 34.51it/s]
Processing recipes by embedding: 100% 98/98 [00:01<00:00, 51.07it/s]
Processing recipes by counting: 100% 98/98 [00:00<00:00, 174.99it/s]
```